+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Ability to browse epics and stories with support for cache-based completion
 - Ability to update the status of stories from the story view buffer
 - Ability to pre-populate the story and epic cache via the Shortcut Search API
+- Completed stories in completion candidates are now displayed with strikethrough and grey color
 
 ### Changed